Move `RGBAColor` out of `WhirlyVector.h`
